requirements.txt
fastapi
uvicorn
requests
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
pytest
httpx
python-dateutil
geopy
